split agenda.js into single files
