+++ index.js
@@ -1,3 +1,3 @@
-const Agenda = require('./lib/agenda.js');
+const Agenda = require('./lib/agenda');
 
 module.exports = Agenda;
