+++ lib/agenda/default-concurrency.js
@@ -1,0 +1,13 @@
+'use strict';
+const debug = require('debug')('agenda:defaultConcurrency');
+
+/**
+ * Set the default concurrency for each job
+ * @param {Number} num default concurrency
+ * @returns {exports} agenda instance
+ */
+module.exports = function(num) {
+  debug('Agenda.defaultConcurrency(%d)', num);
+  this._defaultConcurrency = num;
+  return this;
+};
