+++ lib/agenda/mongo.js
@@ -1,0 +1,14 @@
+'use strict';
+
+/**
+ * Build method used to add MongoDB connection details
+ * @param {MongoClient} mdb instance of MongoClient to use
+ * @param {String} collection name collection we want to use ('agendaJobs')
+ * @param {Function} cb called when MongoDB connection fails or passes
+ * @returns {exports} instance of Agenda
+ */
+module.exports = function(mdb, collection, cb) {
+  this._mdb = mdb;
+  this.db_init(collection, cb);
+  return this;
+};
