+++ lib/agenda/name.js
@@ -1,0 +1,13 @@
+'use strict';
+const debug = require('debug')('agenda:name');
+
+/**
+ * Set name of queue
+ * @param {String} name name of agenda instance
+ * @returns {exports} agenda instance
+ */
+module.exports = function(name) {
+  debug('Agenda.name(%s)', name);
+  this._name = name;
+  return this;
+};
