+++ lib/job.js
@@ -48,7 +48,7 @@ const Job = function(args) {
   }
 
   // Set defaults if undefined
-  // NOTE: What is the difference between 'once' here and 'single' in agenda.js?
+  // NOTE: What is the difference between 'once' here and 'single' in agenda/index.js?
   attrs.nextRunAt = attrs.nextRunAt || new Date();
   attrs.type = attrs.type || 'once';
   this.attrs = attrs;
